requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
boto3==1.29.6
botocore==1.32.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.104.1
fastapi-directory-routing==0.0.1
greenlet==3.0.1
h11==0.14.0
idna==3.4
jmespath==1.0.1
numpy==1.26.2
opencv-python==4.8.1.78
psycopg2-binary==2.9.9
pyasn1==0.5.1
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.6
pyzbar==0.1.9
requests==2.31.0
rsa==4.9
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.24.0.post1
